requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
beautifulsoup4==4.13.3
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
bs4==0.0.2
celery==5.5.3
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.4
curl_cffi==0.9.0
daphne==4.2.0
Django==5.2.1
django-debug-toolbar==5.2.0
django-dirtyfields==1.9.7
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-yasg==1.21.10
Flask==3.1.0
frozenlist==1.5.0
g4f==0.4.7.5
gevent==25.5.1
greenlet==3.2.3
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.5
kombu==5.5.4
loguru==0.7.3
lxml==5.3.0
MarkupSafe==3.0.2
msgpack==1.1.1
multidict==6.1.0
nest-asyncio==1.6.0
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
propcache==0.3.0
psycopg==3.2.7
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.9.0
pyOpenSSL==25.1.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.3
rsa==4.9.1
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
win32_setctime==1.2.0
yarl==1.18.3
zope.event==5.0
zope.interface==7.2
